requirements.txt
matplotlib==3.9.2
numpy==2.1.2
Pillow==11.0.0
scikit_learn==1.5.2
seaborn==0.13.2
streamlit==1.39.0
torch==2.5.0+cu124
torchvision==0.20.0
tqdm==4.66.5